requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
crispy-bootstrap4==2023.1
Django==4.2.6
django-crispy-forms==2.1
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
